Fixed a lot of compilation errors
- Aligned Allocator would not compile
- Loader errors
- Semi-fxed error with macros in Error (needs testing)

+++ src/RawMaterial.cpp
@@ -157,7 +157,7 @@ const std::vector<RawTexture2D::RawTextureData*>& RawMaterial::getDiffuseTexture
 
 unsigned int RawMaterial::getNumOfDiffuseTextures() const
 {
-	return _diffuseTextures.size();
+	return static_cast<unsigned int>(_diffuseTextures.size());
 }
